ipsec: fix offsets and length checks when NULL encryption
The location of the protocol number of the encapsulated data was hard
coded to 14 bytes from the end of the packet.  Null encryption has no
IV or authentication data.  Fix the offsets so that the inner packets
can be decoded.

Also select the correct areas in the hex dump for the ESP pad length
and next protocol number.